Add arguments to set the desired resolution.
This adds the --desired-width and --desired-height arguments which
tell kmscon what resolution the user *desires*. They are specifically
called desired because it is not guaranteed that this resolution will
be used: it will only work if the exact width and height match a
supported mode as reported by the system, and it will only work in
DRM mode (kmscon does not generally support modesetting in fbdev).

Additionally, it might not be possible to use the mode if there is not
enough memory to store the framebuffers. Using `video=<width>x<height>` on
the kernel's command line was enough to make sure that sufficient memory
was available on my machine. In this case, if we try to use the desired
mode then the user will be presented with a blank screen.

In either of the above cases, kmscon will revert to using the default
mode, which is what kmscon always uses prior to this patch.